pyproject.toml
[tool.poetry]
name = "new_admin_panel_sprint_3"
version = "0.1.0"
description = "Project Task: ES ETL service"
authors = ["rgdnk <ruslan.godynyuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.app.dependencies]
django-admin = "^2.0.2"
django = "4.2.5"
python-dotenv = "^1.0.1"
psycopg = "^3.1.17"
psycopg2-binary = "^2.9.9"
django-split-settings = "^1.2.0"
django-debug-toolbar = "^4.2.0"
psycopg2 = "^2.9.9"
tenacity = "^8.2.3"
elasticsearch = "^8.12.1"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
redis = "^5.0.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
mypy = "^1.8.0"
flake8 = "5.0.4"
flake8-blind-except = "0.2.1"
flake8-broken-line = "0.6.0"
flake8-bugbear = "22.9.23"
flake8-builtins = "2.0.0"
flake8-class-attributes-order = "0.1.3"
flake8-cognitive-complexity = "0.1.0"
flake8-commas = "2.1.0"
flake8-comprehensions = "3.10.0"
flake8-debugger = "4.1.2"
flake8-eradicate = "1.4.0"
flake8-functions = "0.0.7"
flake8-isort = "5.0.0"
flake8-mutable = "1.2.0"
flake8-print = "5.0.0"
flake8-pytest = "1.4"
flake8-pytest-style = "1.6.0"
flake8-quotes = "3.3.1"
flake8-string-format = "0.3.0"
flake8-variables-names = "0.0.5"

[tool.poetry.group.etc.dependencies]
numpy = "^1.26.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"
pre-commit = "^3.6.2"
rich = "^13.7.1"

[tool.isort]
include_trailing_comma = true
use_parentheses = true
line_length = 120
multi_line_output = 3
ensure_newline_before_comments = true
force_grid_wrap = 2
lines_before_imports = 1
lines_after_imports = 2
order_by_type = true
group_by_package = true
import_heading_future = "Future"
import_heading_stdlib = "Standard Library"
import_heading_firstparty = "First Party"
import_heading_thirdparty = "Third Party"
import_heading_localfolder = "Local Folder"
known_first_party = [
        "config",
        "movies",
        "v1",
        "locale",
        "postgres_to_es",
        "etl",
    ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
